tools/opai-bx4/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
python-jose[cryptography]
python-multipart
anthropic
